refactor, structuring more correctly
