+++ java/core/src/main/java/co/worklytics/psoxy/storage/impl/ColumnarBulkDataSanitizerImpl.java
@@ -9,6 +9,8 @@ import com.avaulta.gateway.pseudonyms.impl.Base64UrlSha256HashPseudonymEncoder;
 import com.avaulta.gateway.pseudonyms.impl.UrlSafeTokenPseudonymEncoder;
 import com.avaulta.gateway.rules.ColumnarRules;
 import co.worklytics.psoxy.storage.BulkDataSanitizer;
+import com.avaulta.gateway.rules.transforms.FieldTransformPipeline;
+import com.avaulta.gateway.rules.transforms.FieldTransform;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.annotations.VisibleForTesting;
@@ -95,7 +97,7 @@ public class ColumnarBulkDataSanitizerImpl implements BulkDataSanitizer {
             asSetWithCaseInsensitiveComparator(rules.getColumnsToPseudonymizeIfPresent());
 
 
-        Map<String, ColumnarRules.FieldTransformPipeline> columnsToTransform =
+        Map<String, FieldTransformPipeline> columnsToTransform =
             Optional.ofNullable(rules.getFieldsToTransform()).orElse(Collections.emptyMap())
             .entrySet().stream()
             .collect(Collectors.toMap(
@@ -180,12 +182,12 @@ public class ColumnarBulkDataSanitizerImpl implements BulkDataSanitizer {
         Map<String, Pseudonymizer> pseudonymizers = new HashMap<>();
         BiFunction<String, String, String> applyTransformIfAny = (originalColumnName, value) -> {
           if (columnsToTransform.containsKey(originalColumnName.toLowerCase())) {
-              ColumnarRules.FieldTransformPipeline pipeline = columnsToTransform.get(originalColumnName.toLowerCase());
+              FieldTransformPipeline pipeline = columnsToTransform.get(originalColumnName.toLowerCase());
 
-              for ( ColumnarRules.FieldValueTransform transform : pipeline.getTransforms()) {
+              for ( FieldTransform transform : pipeline.getTransforms()) {
                   if (value != null) {
-                      if (transform instanceof ColumnarRules.FieldValueTransform.Filter) {
-                          Pattern pattern = Pattern.compile(((ColumnarRules.FieldValueTransform.Filter) transform).getFilter());
+                      if (transform instanceof FieldTransform.Filter) {
+                          Pattern pattern = Pattern.compile(((FieldTransform.Filter) transform).getFilter());
                           Matcher matcher = pattern.matcher(value);
                           if (matcher.matches()) {
                               if (matcher.groupCount() > 0) {
@@ -196,15 +198,15 @@ public class ColumnarBulkDataSanitizerImpl implements BulkDataSanitizer {
                           }
                       }
 
-                      if (transform instanceof ColumnarRules.FieldValueTransform.FormatString) {
-                          value = String.format(((ColumnarRules.FieldValueTransform.FormatString) transform).getFormatString(), value);
+                      if (transform instanceof FieldTransform.FormatString) {
+                          value = String.format(((FieldTransform.FormatString) transform).getFormatString(), value);
                       }
 
-                      if (transform instanceof ColumnarRules.FieldValueTransform.PseudonymizeWithScope) {
+                      if (transform instanceof FieldTransform.PseudonymizeWithScope) {
                           Pseudonymizer scopedPseudonymizer = pseudonymizer;
                           if (pseudonymizer.getOptions().getPseudonymImplementation() == PseudonymImplementation.LEGACY) {
                               scopedPseudonymizer = pseudonymizers.computeIfAbsent(
-                                  ((ColumnarRules.FieldValueTransform.PseudonymizeWithScope) transform).getPseudonymizeWithScope(),
+                                  ((FieldTransform.PseudonymizeWithScope) transform).getPseudonymizeWithScope(),
                                   scope -> pseudonymizerImplFactory.create(pseudonymizer.getOptions().withDefaultScopeId(scope)));
                           }
                           value = pseudonymizationFunction.apply(value, pipeline.getNewName(), scopedPseudonymizer);
